nn files update

+++ python/needle/nn/nn_basic.py
@@ -282,7 +282,7 @@ class Dropout(Module):
     def forward(self, x: Tensor) -> Tensor:
         ### BEGIN YOUR SOLUTION
         if self.training:
-            mask = init.randb(*x.shape, p=1-self.p)
+            mask = init.randb(*x.shape, p=1-self.p, dtype="float32")
             mask = mask / (1 - self.p)
             return x * mask
         else:
